refactor(test): renamed get_models_to_test.py to get_data_to_test.py
