+++ src/tests/utils/__init__.py
@@ -1,3 +1,3 @@
-from .get_models_to_test import get_models_to_test
+from .get_data_to_test import get_models_to_test, get_inputs_to_test
 
-__all__ = ["get_models_to_test"]
+__all__ = ["get_models_to_test", "get_inputs_to_test"]
